requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.1
django-allauth==0.47.0
django-autofixture==0.12.1
django-cors-headers==3.11.0
django-filter==21.1
django-livesync==0.5
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-seed==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-writable-nested==0.6.3
factory-boy==3.2.1
Faker==12.1.0
filelock==3.4.2
idna==3.3
importlib-metadata==4.10.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
mysqlclient==2.1.0
oauthlib==3.2.0
openapi-codec==1.3.2
Pillow==9.0.0
pipenv==2022.1.8
platformdirs==2.4.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2

toml==0.10.2
toposort==1.7
tornado==6.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.0
virtualenv-clone==0.5.7
zipp==3.7.0
watchdog==2.1.6
websocket-client==1.2.3
